requirements.txt
discord.py>=2.3.0
openai>=1.30.0
motor>=3.3.0
pymongo>=4.7.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.0
itsdangerous>=2.2.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
razorpay>=1.4.0
psutil>=5.9.0
certifi>=2024.0.0
ruff>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-multipart>=0.0.9
wavelink>=3.2.0
yt-dlp>=2024.1.0
Pillow>=10.0.0
deep-translator>=1.11.0
transformers>=4.40.0
torch>=2.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0
